move ArtifactFileReference class to internal.io.reference package
