+++ components/uuf-core/src/main/java/org/wso2/carbon/uuf/internal/io/ArtifactThemeReference.java
@@ -22,6 +22,7 @@ import org.wso2.carbon.uuf.api.reference.FileReference;
 import org.wso2.carbon.uuf.api.reference.ThemeReference;
 import org.wso2.carbon.uuf.exception.FileOperationException;
 import org.wso2.carbon.uuf.internal.io.reference.ArtifactAppReference;
+import org.wso2.carbon.uuf.internal.io.reference.ArtifactFileReference;
 
 import java.nio.file.Files;
 import java.nio.file.Path;
